+++ sdk/core/azure-core/test/ut/service_tracing_test.cpp
@@ -48,11 +48,12 @@ TEST(DiagnosticTracingFactory, SimpleServiceSpanTests)
     Azure::Core::Tracing::_internal::DiagnosticTracingFactory serviceTrace(
         clientOptions, "my-service-cpp", "1.0b2");
 
-    auto contextAndSpan = serviceTrace.CreateSpan("My API", {});
+    auto contextAndSpan = serviceTrace.CreateSpan(
+        "My API", Azure::Core::Tracing::_internal::SpanKind::Internal, {});
     EXPECT_FALSE(contextAndSpan.first.IsCancelled());
   }
 }
-
+namespace {
 // Dummy service tracing class.
 class TestSpan final : public Azure::Core::Tracing::_internal::Span {
 public:
@@ -60,6 +61,7 @@ public:
 
   // Inherited via Span
   virtual void AddAttributes(AttributeSet const&) override {}
+  virtual void AddAttribute(std::string const&, std::string const&) override {}
   virtual void AddEvent(std::string const&, AttributeSet const&) override {}
   virtual void AddEvent(std::string const&) override {}
   virtual void AddEvent(std::exception const&) override {}
@@ -67,6 +69,9 @@ public:
 
   // Inherited via Span
   virtual void End(Azure::Nullable<Azure::DateTime>) override {}
+
+  // Inherited via Span
+  virtual void PropagateToHttpHeaders(Azure::Core::Http::Request&) override {}
 };
 
 class TestAttributeSet : public Azure::Core::Tracing::_internal::AttributeSet {
@@ -107,7 +112,7 @@ public:
     return std::make_shared<TestTracer>(serviceName, serviceVersion);
   };
 };
-
+} // namespace
 TEST(DiagnosticTracingFactory, BasicServiceSpanTests)
 {
   {
@@ -115,14 +120,16 @@ TEST(DiagnosticTracingFactory, BasicServiceSpanTests)
     Azure::Core::Tracing::_internal::DiagnosticTracingFactory serviceTrace(
         clientOptions, "my-service-cpp", "1.0b2");
 
-    auto contextAndSpan = serviceTrace.CreateSpan("My API", {});
-    auto span = std::move(contextAndSpan.second);
+    auto contextAndSpan = serviceTrace.CreateSpan(
+        "My API", Azure::Core::Tracing::_internal::SpanKind::Internal, {});
+    ServiceSpan span = std::move(contextAndSpan.second);
 
     span.End();
     span.AddEvent("New Event");
     span.AddEvent(std::runtime_error("Exception"));
     span.SetStatus(SpanStatus::Error);
   }
+
   {
     Azure::Core::_internal::ClientOptions clientOptions;
     auto testTracer = std::make_shared<TestTracingProvider>();
@@ -130,8 +137,9 @@ TEST(DiagnosticTracingFactory, BasicServiceSpanTests)
     Azure::Core::Tracing::_internal::DiagnosticTracingFactory serviceTrace(
         clientOptions, "my-service-cpp", "1.0b2");
 
-    auto contextAndSpan = serviceTrace.CreateSpan("My API", {});
-    auto span = std::move(contextAndSpan.second);
+    auto contextAndSpan = serviceTrace.CreateSpan(
+        "My API", Azure::Core::Tracing::_internal::SpanKind::Internal, {});
+    ServiceSpan span = std::move(contextAndSpan.second);
 
     span.End();
     span.AddEvent("New Event");